Replace operator::Operator* by operator::LinearOperator
