+++ src/krylov/mod.rs
@@ -6,7 +6,6 @@ use ndarray::*;
 pub mod arnoldi;
 pub mod householder;
 pub mod mgs;
-pub mod operator;
 
 pub use arnoldi::{arnoldi_householder, arnoldi_mgs, Arnoldi};
 pub use householder::{householder, Householder};
