+++ src/operator.rs
@@ -1,105 +1,80 @@
-//! Linear Operator
+//! Linear operator algebra
 
+use crate::generate::hstack;
+use crate::types::*;
 use ndarray::*;
 
-use super::types::*;
+pub trait LinearOperator {
+    type Elem: Scalar;
 
-pub trait Operator<A, S, D>
-where
-    S: Data<Elem = A>,
-    D: Dimension,
-{
-    fn op(&self, a: &ArrayBase<S, D>) -> Array<A, D>;
-}
-
-pub trait OperatorInto<S, D>
-where
-    S: DataMut,
-    D: Dimension,
-{
-    fn op_into(&self, a: ArrayBase<S, D>) -> ArrayBase<S, D>;
-}
-
-pub trait OperatorInplace<S, D>
-where
-    S: DataMut,
-    D: Dimension,
-{
-    fn op_inplace<'a>(&self, a: &'a mut ArrayBase<S, D>) -> &'a mut ArrayBase<S, D>;
-}
-
-impl<T, A, S, D> Operator<A, S, D> for T
-where
-    A: Scalar + Lapack,
-    S: Data<Elem = A>,
-    D: Dimension,
-    T: linalg::Dot<ArrayBase<S, D>, Output = Array<A, D>>,
-{
-    fn op(&self, rhs: &ArrayBase<S, D>) -> Array<A, D> {
-        self.dot(rhs)
-    }
-}
-
-pub trait OperatorMulti<A, S, D>
-where
-    S: Data<Elem = A>,
-    D: Dimension,
-{
-    fn op_multi(&self, a: &ArrayBase<S, D>) -> Array<A, D>;
-}
-
-impl<T, A, S, D> OperatorMulti<A, S, D> for T
-where
-    A: Scalar + Lapack,
-    S: DataMut<Elem = A>,
-    D: Dimension + RemoveAxis,
-    for<'a> T: OperatorInplace<ViewRepr<&'a mut A>, D::Smaller>,
-{
-    fn op_multi(&self, a: &ArrayBase<S, D>) -> Array<A, D> {
-        let a = a.to_owned();
-        self.op_multi_into(a)
-    }
-}
-
-pub trait OperatorMultiInto<S, D>
-where
-    S: DataMut,
-    D: Dimension,
-{
-    fn op_multi_into(&self, a: ArrayBase<S, D>) -> ArrayBase<S, D>;
-}
-
-impl<T, A, S, D> OperatorMultiInto<S, D> for T
-where
-    S: DataMut<Elem = A>,
-    D: Dimension + RemoveAxis,
-    for<'a> T: OperatorInplace<ViewRepr<&'a mut A>, D::Smaller>,
-{
-    fn op_multi_into(&self, mut a: ArrayBase<S, D>) -> ArrayBase<S, D> {
-        self.op_multi_inplace(&mut a);
+    /// Apply operator out-place
+    fn apply<S>(&self, a: &ArrayBase<S, Ix1>) -> Array1<S::Elem>
+    where
+        S: Data<Elem = Self::Elem>,
+    {
+        let mut a = a.to_owned();
+        self.apply_mut(&mut a);
         a
     }
-}
 
-pub trait OperatorMultiInplace<S, D>
-where
-    S: DataMut,
-    D: Dimension,
-{
-    fn op_multi_inplace<'a>(&self, a: &'a mut ArrayBase<S, D>) -> &'a mut ArrayBase<S, D>;
-}
+    /// Apply operator in-place
+    fn apply_mut<S>(&self, a: &mut ArrayBase<S, Ix1>)
+    where
+        S: DataMut<Elem = Self::Elem>,
+    {
+        let b = self.apply(a);
+        azip!(mut a(a), b in { *a = b });
+    }
 
-impl<T, A, S, D> OperatorMultiInplace<S, D> for T
-where
-    S: DataMut<Elem = A>,
-    D: Dimension + RemoveAxis,
-    for<'a> T: OperatorInplace<ViewRepr<&'a mut A>, D::Smaller>,
-{
-    fn op_multi_inplace<'a>(&self, a: &'a mut ArrayBase<S, D>) -> &'a mut ArrayBase<S, D> {
-        let n = a.ndim();
-        for mut col in a.axis_iter_mut(Axis(n - 1)) {
-            self.op_inplace(&mut col);
+    /// Apply operator with move
+    fn apply_into<S>(&self, mut a: ArrayBase<S, Ix1>) -> ArrayBase<S, Ix1>
+    where
+        S: DataOwned<Elem = Self::Elem> + DataMut,
+    {
+        self.apply_mut(&mut a);
+        a
+    }
+
+    /// Apply operator to matrix out-place
+    fn apply2<S>(&self, a: &ArrayBase<S, Ix2>) -> Array2<S::Elem>
+    where
+        S: Data<Elem = Self::Elem>,
+    {
+        let cols: Vec<_> = a.axis_iter(Axis(1)).map(|col| self.apply(&col)).collect();
+        hstack(&cols).unwrap()
+    }
+
+    /// Apply operator to matrix in-place
+    fn apply2_mut<S>(&self, a: &mut ArrayBase<S, Ix2>)
+    where
+        S: DataMut<Elem = Self::Elem>,
+    {
+        for mut col in a.axis_iter_mut(Axis(1)) {
+            self.apply_mut(&mut col)
         }
+    }
+
+    /// Apply operator to matrix with move
+    fn apply2_into<S>(&self, mut a: ArrayBase<S, Ix2>) -> ArrayBase<S, Ix2>
+    where
+        S: DataOwned<Elem = Self::Elem> + DataMut,
+    {
+        self.apply2_mut(&mut a);
         a
+    }
+}
+
+impl<A, Sa> LinearOperator for ArrayBase<Sa, Ix2>
+where
+    A: Scalar,
+    Sa: Data<Elem = A>,
+{
+    type Elem = A;
+
+    fn apply<S>(&self, a: &ArrayBase<S, Ix1>) -> Array1<A>
+    where
+        S: Data<Elem = A>,
+    {
+        self.dot(a)
     }
 }
